PP2 Update

+++ portal/import_template.php
@@ -25,24 +25,40 @@
 require_once("../interface/globals.php");
 
 if ($_POST['mode'] == 'get') {
-    echo file_get_contents($_POST['docid']);
-    exit;
+    if (validateFile($_POST['docid'])) {
+        echo file_get_contents($_POST['docid']);
+        exit();
+    } else {
+        die(xlt('Invalid File'));
+    }
 } else if ($_POST['mode'] == 'save') {
-    file_put_contents($_POST['docid'], $_POST['content']);
-    exit(true);
+    if (validateFile($_POST['docid'])) {
+        if (stripos($_POST['content'], "<?php") !== false) {
+            file_put_contents($_POST['docid'], $_POST['content']);
+            exit(true);
+        } else {
+            die(xlt('Invalid Content'));
+        }
+    } else {
+        die(xlt('Invalid File'));
+    }
 } else if ($_POST['mode'] == 'delete') {
-    unlink($_POST['docid']);
-    exit(true);
+    if (validateFile($_POST['docid'])) {
+        unlink($_POST['docid']);
+        exit(true);
+    } else {
+        die(xlt('Invalid File'));
+    }
 }
 
 // so it is an import
 if (!isset($_POST['up_dir'])) {
-    define("UPLOAD_DIR", $GLOBALS['OE_SITE_DIR'] .  '/documents/onsite_portal_documents/templates/');
+    define("UPLOAD_DIR", $GLOBALS['OE_SITE_DIR'] . '/documents/onsite_portal_documents/templates/');
 } else {
     if ($_POST['up_dir'] > 0) {
-        define("UPLOAD_DIR", $GLOBALS['OE_SITE_DIR'] .  '/documents/onsite_portal_documents/templates/'. $_POST['up_dir'] . '/');
+        define("UPLOAD_DIR", $GLOBALS['OE_SITE_DIR'] . '/documents/onsite_portal_documents/templates/' . $_POST['up_dir'] . '/');
     } else {
-        define("UPLOAD_DIR", $GLOBALS['OE_SITE_DIR'] .  '/documents/onsite_portal_documents/templates/');
+        define("UPLOAD_DIR", $GLOBALS['OE_SITE_DIR'] . '/documents/onsite_portal_documents/templates/');
     }
 }
 
@@ -51,29 +67,47 @@ if (!empty($_FILES["tplFile"])) {
 
     if ($tplFile["error"] !== UPLOAD_ERR_OK) {
         header("refresh:2;url= import_template_ui.php");
-        echo "<p>". xlt("An error occurred: Missing file to upload: Use back button!") . "</p>";
+        echo "<p>" . xlt("An error occurred: Missing file to upload: Use back button!") . "</p>";
         exit;
     }
 
     // ensure a safe filename
     $name = preg_replace("/[^A-Z0-9._-]/i", "_", $tplFile["name"]);
+    if (preg_match("/(.*)\.(php|php3|php4|php5|php7)$/i", $name) !== 0) {
+        die(xlt('Executables not allowed'));
+    }
     $parts = pathinfo($name);
-    $name = $parts["filename"].'.tpl';
+    $name = $parts["filename"] . '.tpl';
     // don't overwrite an existing file
     while (file_exists(UPLOAD_DIR . $name)) {
         $i = rand(0, 128);
-        $newname = $parts["filename"] . "-" . $i . "." . $parts["extension"].".replaced";
-        rename(UPLOAD_DIR .$name, UPLOAD_DIR .$newname);
+        $newname = $parts["filename"] . "-" . $i . "." . $parts["extension"] . ".replaced";
+        rename(UPLOAD_DIR . $name, UPLOAD_DIR . $newname);
     }
 
     // preserve file from temporary directory
     $success = move_uploaded_file($tplFile["tmp_name"], UPLOAD_DIR . $name);
     if (!$success) {
-        echo "<p>". xlt("Unable to save file: Use back button!") . "</p>";
+        echo "<p>" . xlt("Unable to save file: Use back button!") . "</p>";
         exit;
     }
 
     // set proper permissions on the new file
     chmod(UPLOAD_DIR . $name, 0644);
     header("location: " . $_SERVER['HTTP_REFERER']);
+}
+
+function validateFile($filename = '')
+{
+    $valid = false;
+    $filePath = $GLOBALS['OE_SITE_DIR'] . '/documents/onsite_portal_documents/templates';
+    if (stripos($filename, $filePath) === false) {
+        return false;
+    }
+    if (preg_match("/(.*)\.(php|php3|php4|php5|php7)$/i", $filename) === 0) {
+        if (preg_match("/(.*)\.(tpl)$/i", $filename) === 1) {
+            $valid = true;
+        }
+    }
+    return $valid;
 }
